Transactions section almost fully implemented. Problems with the JSPs have to be solved

+++ src/main/java/co/edu/unbosque/entity/BookLot.java
@@ -60,7 +60,7 @@ public class BookLot implements TransactionManager {
 	private Long pricePerUnit;
 
 	@Column(name = "available_units", nullable = false)
-	private Integer availableUnits;
+	private Long availableUnits;
 
 	@JsonIgnore
 	@ManyToOne(cascade = {CascadeType.MERGE, CascadeType.REFRESH})
