+++ src/main/java/co/edu/unbosque/entity/DiscLot.java
@@ -60,7 +60,7 @@ public class DiscLot implements TransactionManager {
 	private Long pricePerUnit;
 
 	@Column(name = "available_units", nullable = false)
-	private Integer availableUnits;
+	private Long availableUnits;
 
 	@JsonIgnore
 	@ManyToOne(cascade = {CascadeType.MERGE, CascadeType.REFRESH})
