+++ src/main/java/co/edu/unbosque/entity/MagazineLot.java
@@ -50,7 +50,7 @@ public class MagazineLot implements TransactionManager {
 	private Long pricePerUnit;
 
 	@Column(name = "available_units", nullable = false)
-	private Integer availableUnits;
+	private Long availableUnits;
 
 	@JsonIgnore
 	@ManyToOne(cascade = {CascadeType.MERGE, CascadeType.REFRESH})
